# specs/api/workspace-sync.openapi.yaml
openapi: 3.1.0
info:
  title: Prodivix Workspace Sync API
  version: 0.2.0
  description: >
    Canonical API for workspace-first editing, partitioned revision recovery,
    and atomic WorkspaceOperation commits. API-003 was amended by the Atomic
    Commit Hard Cut on 2026-07-12; the former /batch contract is removed.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/workspaces/{workspaceId}/documents/{documentId}:
    patch:
      summary: Patch one document with a command
      description: >
        Apply one versioned CommandEnvelope to a Workspace document. The server
        applies command.forwardOps, verifies command.reverseOps can restore the
        previous document, runs document-type validation, advances contentRev,
        and records the command in the operation log. PIR, code, and generic
        documents use different patch-path policies. Only the workspace owner
        may mutate the document; owner mismatches return the same metadata-free
        404 envelope as a missing workspace before conflict details are built.
      operationId: patchDocument
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: string
        - in: path
          name: documentId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchDocumentRequest'
      responses:
        '200':
          description: Patched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MutationSuccessResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Workspace not found or not owned by the caller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Invalid command, forbidden patch path, PIR validation failure, or revision range/capacity failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Revision conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceConflictErrorEnvelope'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque session token
  schemas:
...
    WorkspaceRouteCanonicalString:
      type: string
      minLength: 1
      pattern: '^(?:\S|\S[\s\S]*\S)(?![\s\S])'
      description: Non-empty route identity with no leading or trailing whitespace.
    WorkspaceRouteManifest:
      type: object
      required: [version, root]
      description: >
        Closed canonical RouteManifest wire document. Unknown fields and null
        placeholders are rejected instead of being persisted or discarded.
        Route ids are globally unique, mount ids are unique, and every modules
        map key must exactly equal its value's moduleId. Document references in
        both the project root and every module root are kind-constrained:
        layoutDocId references pir-layout, while pageDocId and outlet binding
        pageDocId reference pir-page or pir-component documents.
      properties:
        version:
          $ref: '#/components/schemas/WorkspaceRouteCanonicalString'
        root:
          $ref: '#/components/schemas/WorkspaceRouteRootNode'
        modules:
          type: object
          propertyNames:
            $ref: '#/components/schemas/WorkspaceRouteCanonicalString'
          additionalProperties:
            $ref: '#/components/schemas/WorkspaceRouteModule'
        mounts:
          type: array
          items:
            $ref: '#/components/schemas/WorkspaceRouteModuleMount'
      additionalProperties: false
    WorkspaceRouteRootNode:
      allOf:
        - $ref: '#/components/schemas/WorkspaceRouteNode'
        - type: object
          properties:
            id:
              type: string
              const: root
    WorkspaceRouteNode:
      type: object
      required: [id]
      properties:
        id:
          $ref: '#/components/schemas/WorkspaceRouteCanonicalString'
        segment:
          type: string
        index:
          type: boolean
        layoutDocId:
          $ref: '#/components/schemas/WorkspaceRouteCanonicalString'
          description: References a pir-layout workspace document.
        pageDocId:
          $ref: '#/components/schemas/WorkspaceRouteCanonicalString'
          description: References a pir-page or pir-component workspace document.
        outletNodeId:
          $ref: '#/components/schemas/WorkspaceRouteCanonicalString'
        outletBindings:
          type: object
          propertyNames:
            $ref: '#/components/schemas/WorkspaceRouteCanonicalString'
          additionalProperties:
            $ref: '#/components/schemas/WorkspaceRouteOutletBinding'
        runtime:
          $ref: '#/components/schemas/WorkspaceRouteRuntime'
        children:
          type: array
          items:
            $ref: '#/components/schemas/WorkspaceRouteNode'
      additionalProperties: false
    WorkspaceRouteOutletBinding:
      type: object
      required: [outletNodeId]
      properties:
        outletNodeId:
          $ref: '#/components/schemas/WorkspaceRouteCanonicalString'
        pageDocId:
          $ref: '#/components/schemas/WorkspaceRouteCanonicalString'
          description: References a pir-page or pir-component workspace document.
      additionalProperties: false
    WorkspaceRouteRuntime:
      type: object
      properties:
        loaderRef:
          $ref: '#/components/schemas/WorkspaceRouteCodeReference'
        actionRef:
          $ref: '#/components/schemas/WorkspaceRouteCodeReference'
        guardRef:
          $ref: '#/components/schemas/WorkspaceRouteCodeReference'
      additionalProperties: false
    WorkspaceRouteCodeReference:
      type: object
      required: [artifactId]
      properties:
        artifactId:
          $ref: '#/components/schemas/WorkspaceRouteCanonicalString'
        exportName:
          $ref: '#/components/schemas/WorkspaceRouteCanonicalString'
        symbolId:
          $ref: '#/components/schemas/WorkspaceRouteCanonicalString'
      additionalProperties: false
    WorkspaceRouteModule:
      type: object
      required: [moduleId, version, root]
      properties:
        moduleId:
          $ref: '#/components/schemas/WorkspaceRouteCanonicalString'
        version:
          $ref: '#/components/schemas/WorkspaceRouteCanonicalString'
        root:
          $ref: '#/components/schemas/WorkspaceRouteNode'
          description: >
            Module route root. Its layoutDocId, pageDocId, and outlet binding
            pageDocId references follow the same document-kind constraints as
            the project route root.
      additionalProperties: false
    WorkspaceRouteModuleMount:
      type: object
      required: [mountId, moduleRef]
      properties:
        mountId:
          $ref: '#/components/schemas/WorkspaceRouteCanonicalString'
        moduleRef:
          $ref: '#/components/schemas/WorkspaceRouteCanonicalString'
        mountPath:
          type: string
        parentRouteNodeId:
          $ref: '#/components/schemas/WorkspaceRouteCanonicalString'
      additionalProperties: false
    WorkspaceTree:
      type: object
      required: [treeRootId, treeById]
      properties:
        treeRootId:
          $ref: '#/components/schemas/WorkspaceOperationCanonicalId'
        treeById:
          type: object
          minProperties: 1
          propertyNames:
            $ref: '#/components/schemas/WorkspaceOperationCanonicalId'
          additionalProperties:
            $ref: '#/components/schemas/WorkspaceTreeNode'
      additionalProperties: false
    WorkspaceTreeNode:
      oneOf:
        - type: object
          required: [id, kind, name, parentId, children]
          properties:
            id:
              $ref: '#/components/schemas/WorkspaceOperationCanonicalId'
            kind:
              type: string
              const: dir
            name:
              type: string
            parentId:
              oneOf:
                - $ref: '#/components/schemas/WorkspaceOperationCanonicalId'
                - type: 'null'
            children:
              type: array
              uniqueItems: true
              items:
                $ref: '#/components/schemas/WorkspaceOperationCanonicalId'
          additionalProperties: false
        - type: object
          required: [id, kind, name, parentId, docId]
          properties:
            id:
              $ref: '#/components/schemas/WorkspaceOperationCanonicalId'
            kind:
              type: string
              const: doc
            name:
              type: string
            parentId:
              $ref: '#/components/schemas/WorkspaceOperationCanonicalId'
            docId:
              $ref: '#/components/schemas/WorkspaceOperationCanonicalId'
          additionalProperties: false
      discriminator:
        propertyName: kind
    WorkspaceDocument:
      type: object
      required: [id, type, path, contentRev, metaRev, content]
      properties:
        id:
          $ref: '#/components/schemas/WorkspaceOperationCanonicalId'
        name:
          type: string
          pattern: '\S'
        type:
          $ref: '#/components/schemas/WorkspaceDocumentType'
        path:
          type: string
        contentRev:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        metaRev:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        content:
          description: JSON document content. Code documents use the language/source wrapper.
        capabilities:
          type: array
          minItems: 1
          uniqueItems: true
          description: >
            Omitted when empty. Otherwise values are non-empty canonical strings
            sorted by Unicode code-point order.
          items:
            $ref: '#/components/schemas/WorkspaceOperationCanonicalId'
        updatedAt:
          type: string
          format: date-time
          description: >
            Optional in shared authoring-state documents. Server mutation
            acknowledgements require the authoritative timestamp.
      additionalProperties: false
    WorkspaceDocumentType:
      type: string
      enum:
        - pir-page
        - pir-layout
        - pir-component
        - pir-graph
        - pir-animation
        - code
        - asset
        - project-config
    PatchDocumentRequest:
      type: object
      required: [expectedContentRev, command]
      properties:
        expectedContentRev:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        command:
          $ref: '#/components/schemas/CommandEnvelope'
        clientMutationId:
          type: string
...
    PatchOp:
      type: object
      required: [op, path]
      description: >
        Restricted JSON Patch operation. PIR documents allow /ui/graph, /logic,
        /animation, /metadata, and /x-*; code documents allow /language, /source,
        /metadata, and /x-*; other document types allow non-root paths. The
        document root path / is always forbidden, and /ui/root is additionally
        forbidden for PIR documents. Array "-" is only valid for add.
      properties:
        op:
          type: string
          enum: [add, remove, replace, move, copy, test]
        path:
          type: string
        from:
          type: string
        value:
          description: Operation payload
      additionalProperties: false
    CommandEnvelope:
      type: object
      required:
        [id, namespace, type, version, issuedAt, forwardOps, reverseOps, target]
      properties:
        id:
          type: string
        namespace:
          type: string
        type:
          type: string
        version:
          type: string
          description: semantic version string, e.g. 1.0
        issuedAt:
          type: string
          format: date-time
        forwardOps:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PatchOp'
        reverseOps:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/PatchOp'
        target:
          type: object
          required: [workspaceId]
          not:
            required: [documentId, routeNodeId]
          properties:
            workspaceId:
              type: string
            documentId:
              type: string
            routeNodeId:
              type: string
          additionalProperties: false
        mergeKey:
          type: string
        label:
          type: string
        domainHint:
          type: string
          enum: [pir, workspace, route, nodegraph, animation, code]
      additionalProperties: false
    WorkspaceOperationCanonicalId:
      type: string
      minLength: 1
      pattern: '^(?:\S|\S[\s\S]*\S)(?![\s\S])'
      description: Non-empty identifier with no leading or trailing whitespace.
...
    MutationSuccessResponse:
      type: object
      required: [workspaceId, workspaceRev, routeRev, opSeq]
      properties:
        workspaceId:
          type: string
        workspaceRev:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        routeRev:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        updatedDocuments:
          type: array
          items:
            $ref: '#/components/schemas/WorkspaceDocument'
        removedDocumentIds:
          type: array
          items:
            type: string
        tree:
          $ref: '#/components/schemas/WorkspaceTree'
        routeManifest:
          $ref: '#/components/schemas/WorkspaceRouteManifest'
        settings:
          type: object
          additionalProperties: true
        opSeq:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        acceptedMutationId:
          type: string
      additionalProperties: false
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorPayload'
    WorkspaceConflictErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          allOf:
            - $ref: '#/components/schemas/ErrorPayload'
            - type: object
              required: [details]
              properties:
                details:
                  $ref: '#/components/schemas/WorkspaceConflictDetails'
      additionalProperties: false
    WorkspaceConflictDetails:
      type: object
      required: [conflictType, workspaceId, expected, current]
      allOf:
        - if:
            properties:
              conflictType:
                const: WORKSPACE_CONFLICT
            required: [conflictType]
          then:
            properties:
              expected:
                allOf:
                  - required: [workspaceRev]
                  - not:
                      required: [document]
              current:
                not:
                  required: [document]
        - if:
            properties:
              conflictType:
                const: ROUTE_CONFLICT
            required: [conflictType]
          then:
            properties:
              expected:
                allOf:
                  - required: [workspaceRev, routeRev]
                  - not:
                      required: [document]
              current:
                not:
                  required: [document]
        - if:
            properties:
              conflictType:
                const: DOCUMENT_CONFLICT
            required: [conflictType]
          then:
            properties:
              expected:
                required: [document]
              current:
                required: [document]
      properties:
        conflictType:
          type: string
          enum: [WORKSPACE_CONFLICT, ROUTE_CONFLICT, DOCUMENT_CONFLICT]
        workspaceId:
          type: string
        expected:
          $ref: '#/components/schemas/WorkspaceConflictExpectedRevisions'
        current:
          $ref: '#/components/schemas/WorkspaceConflictCurrentRevisions'
      additionalProperties: false
    WorkspaceConflictExpectedRevisions:
      type: object
      minProperties: 1
      anyOf:
        - required: [workspaceRev]
        - required: [routeRev]
        - required: [document]
      properties:
        workspaceRev:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        routeRev:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        document:
          $ref: '#/components/schemas/WorkspaceConflictExpectedDocument'
      additionalProperties: false
    WorkspaceConflictExpectedDocument:
      type: object
      required: [id]
      description: >
        At least one revision property is required. Metadata-only conflicts
        still use DOCUMENT_CONFLICT. Both explicit nulls encode an absence
        precondition for concurrent document creation.
      anyOf:
        - required: [contentRev]
        - required: [metaRev]
      allOf:
        - if:
            required: [contentRev]
            properties:
              contentRev:
                type: 'null'
          then:
            required: [metaRev]
            properties:
              metaRev:
                type: 'null'
        - if:
            required: [metaRev]
            properties:
              metaRev:
                type: 'null'
          then:
            required: [contentRev]
            properties:
              contentRev:
                type: 'null'
      properties:
        id:
          type: string
        contentRev:
          type: [integer, 'null']
          minimum: 1
          maximum: 9007199254740991
        metaRev:
          type: [integer, 'null']
          minimum: 1
          maximum: 9007199254740991
      additionalProperties: false
    WorkspaceConflictCurrentRevisions:
      type: object
      required: [workspaceRev, routeRev, opSeq]
      properties:
        workspaceRev:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        routeRev:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        opSeq:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        document:
          $ref: '#/components/schemas/WorkspaceConflictCurrentDocument'
      additionalProperties: false
    WorkspaceConflictCurrentDocument:
      oneOf:
        - $ref: '#/components/schemas/WorkspaceConflictCurrentDocumentMetadata'
        - type: 'null'
      description: >
        Safe current metadata, or null when the expected document was deleted
        remotely. The document property remains required for DOCUMENT_CONFLICT.
    WorkspaceConflictCurrentDocumentMetadata:
      type: object
      required: [id, type, path, contentRev, metaRev, updatedAt]
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/WorkspaceDocumentType'
        path:
          type: string
        contentRev:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        metaRev:
          type: integer
          minimum: 1
          maximum: 9007199254740991
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
    ErrorPayload:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          examples:
            - WKS-4001
            - WKS-4002
            - WKS-4003
            - WKS-5001
            - WKS-5002
            - PIR-4001
            - API-1001
        severity:
          type: string
          enum: [info, warning, error, fatal]
        domain:
          type: string
        retryable:
          type: boolean
        requestId:
          type: string
        docsUrl:
          type: string
        message:
          type: string
        details:
          description: Safe machine-readable context. Revision conflicts use WorkspaceConflictDetails and never include document content.
        diagnostics:
          type: array
          items:
            $ref: '#/components/schemas/BackendDiagnostic'
    BackendDiagnostic:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [info, warning, error, fatal]
        domain:
          type: string
        retryable:
          type: boolean
        path:
          type: string
        targetRef:
          type: object
          additionalProperties: true
        docsUrl:
          type: string
        details:
          additionalProperties: true
      additionalProperties: true